requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-telegram-bot==21.0.1
qrcode[pil]==7.4.2
python-dotenv==1.0.1
pydantic==2.6.4